+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # OSIM Changelog
 
+## [Unreleased]
+### Added
+* Allow grouping affects by ps_module (`OSIDB-4546`)
+
 ## [2025.10.0]
 ### Fixed
 * Fix multi-row operations in affects table (`OSIDB-4381`)
@@ -20,6 +24,7 @@
   * Allow sorting by any column with the option to combine them
   * Added bulk actions for filing trackers
   * Fixed some issues of old implementation
+  * Allow grouping affects by PsModule (`OSIDB-4546`)
 
 
 ## [2025.9.0]